+++ tests/src/imap/fetch.rs
@@ -71,18 +71,22 @@ pub async fn test(imap: &mut ImapConnection, _imap_check: &mut ImapConnection) {
     .await;
     imap.assert_read(Type::Tagged, ResponseType::Ok)
         .await
-        .assert_contains("BINARY[1] {175}")
+        .assert_contains("BINARY[1] ~{175}")
         .assert_contains("BINARY.SIZE[1] 175")
         .assert_contains("BODY[1.TEXT] {239}")
         .assert_contains("BODY[2.1.HEADER] {88}")
-        .assert_contains("BINARY[2.1] {101}")
+        .assert_contains("BINARY[2.1] ~{108}")
         .assert_contains("BODY[MIME] {54}")
         .assert_contains("BODY[HEADER.FIELDS (FROM)]<10> {8}")
         .assert_contains("&ldquo;exporting&rdquo;")
         .assert_contains("PGh0bWw+PHA+")
         .assert_contains("Content-Transfer-Encoding: quoted-printable")
-        .assert_contains("ℌ𝔢𝔩𝔭 𝔪𝔢 𝔢𝔵𝔭𝔬𝔯𝔱 𝔪𝔶 𝔟𝔬𝔬𝔨")
         .assert_contains("Vandelay");
+    let fraktur_utf16_le: Vec<u8> = "ℌ𝔢𝔩𝔭 𝔪𝔢 𝔢𝔵𝔭𝔬𝔯𝔱 𝔪𝔶 𝔟𝔬𝔬𝔨"
+        .encode_utf16()
+        .flat_map(|c| c.to_le_bytes())
+        .collect();
+    imap.assert_last_contains_bytes(&fraktur_utf16_le);
 
     // We are in EXAMINE mode, fetching body should not set \Seen
     imap.send("UID FETCH 10 (FLAGS)").await;
@@ -99,7 +103,7 @@ pub async fn test(imap: &mut ImapConnection, _imap_check: &mut ImapConnection) {
         .await;
     imap.assert_read(Type::Tagged, ResponseType::Ok)
         .await
-        .assert_contains("BINARY[1] {175}")
+        .assert_contains("BINARY[1] ~{175}")
         .assert_contains("BINARY.SIZE[1] 175")
         .assert_contains("BODY[1.TEXT] {239}");